pyproject.toml
[tool.poetry]
name = "Liteyuki"
version = "1.0.0"
description = "基于NoneBot2的QQ机器人"
authors = ["SnowyKami","TimeRainStarSky"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"
nb-cli = "^0.6.8"
nonebot2 = "^2.0.0rc2"
nonebot-adapter-onebot = "^2.2.0"
pymongo = "^4.3.3"
pillow = "^9.3.0"
aiohttp = "^3.8.3"
psutil = "^5.9.1"
jieba = "~0.42.1"
[tool.nonebot]
plugins = []
plugin_dirs = ["src/nonebot_plugins", "src/liteyuki_plugins"]

[[tool.poetry.source]]
name = "Tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true